Python Programming
In the E-Commerce shopping experience, it is essential to deliver the product promptly to the customer. The courier company stores the relevant data of its clients and parcels in the form of a python dictionary and different analyses were performed on the data on the delivery performance using dictionary comprehensions.