Numpy Pandas  and Visualization
The dataset used is about medical expenditure of all the employees for the year 2019. The analysis is performed to understand the relation between medical expenditure and health parameters and other data like a number of dependents, age, bmi, smoking and alcoholic habits. Investigations were performed on the effect of age, smoking/alcohol behaviors on health expenditures.