Exploratory Data Analysis
Dataset contains the loan listings of p2p (peer to peer loan) lending platform called “Prosper”. It connects borrowers with the lenders and “Prosper” (platform provider) & handles servicing of the loan on behalf of the matched borrowers and investors. This project provides insights to the borrowers & lenders to make decisions to invest on the respective loan listing. The answers were obtained by exploring the data to understand the relations amongst the variables with appropriate visualization