Database Management Systems - I
To learn the insights of the complete match details and about the players, which helps the bidder to bid the IPL matches. To facilitate the projections of the winner at the end of every match and update the team standings in the tournament along with the bidder points table. To visualize the leader board rankings of top 3 positions at any given point of time.