Unsupervised Learning
Online shopping has changed the landscape of buying products and it offers immense flexibility to the end customer. Generally, in an online retail store, the businesses are interested in knowing the purchase behavior so that they can customize their web pages based on user preferences. This makes online shopping a wonderful experience. Here, in this case study you are expected to mine patterns using Clustering Techniques and see if it has an impact on the revenue of the business or not.